refactor(Kitchenpos): refactor Menu Aggregate
1. Entity 연관관계 설정
2. getter 제거
