--- src/main/java/kitchenpos/menus/tobe/v2/domain/MenuProduct.java
@@ -1,60 +1,0 @@
-package kitchenpos.menus.tobe.v2.domain;
-
-
-import com.fasterxml.jackson.annotation.JsonIgnore;
-import kitchenpos.products.tobe.domain.Product;
-
-import javax.persistence.*;
-import java.math.BigDecimal;
-
-@Entity
-public class MenuProduct {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long seq;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "menu_id")
-    @JsonIgnore
-    private Menu menu;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "product_id")
-    private Product product;
-
-    private long quantity;
-
-    public MenuProduct() {
-    }
-
-    public MenuProduct(Product product, long quantity) {
-        this.product = product;
-        this.quantity = quantity;
-    }
-
-    @JsonIgnore
-    public BigDecimal getTotalPrice() {
-        return product.getPrice().multiply(BigDecimal.valueOf(quantity));
-    }
-
-    public void changeMenu(Menu menu) {
-        this.menu = menu;
-        menu.getMenuProducts().add(this);
-    }
-
-    public Long getSeq() {
-        return seq;
-    }
-
-    public Menu getMenu() {
-        return menu;
-    }
-
-    public Product getProduct() {
-        return product;
-    }
-
-    public long getQuantity() {
-        return quantity;
-    }
-}
